+++ test/externals/ruby_test/test/core/Dir/class/tc_mkdir.rb
@@ -10,8 +10,8 @@ class TC_Dir_Mkdir_Class < Test::Unit::TestCase
    include Test::Helper
 
    def setup
-      @dir1 = "bogus1"
-      @dir2 = "bogus2"
+      @dir1 = File.join(base_dir(__FILE__), 'mkdir_test1')
+      @dir2 = File.join(base_dir(__FILE__), 'mkdir_test2')
    end
 
    def test_mkdir_basic
@@ -22,9 +22,14 @@ class TC_Dir_Mkdir_Class < Test::Unit::TestCase
 
    def test_mkdir
       assert_equal(0, Dir.mkdir(@dir1))
+      assert(File.exists?(@dir1))
+   end
+
+   def test_mkdir_with_mode
       assert_equal(0, Dir.mkdir(@dir2, 0777))
+      assert(File.exists?(@dir2))
       unless WINDOWS || JRUBY
-         assert_equal(0, File.umask & File.stat(@dir1).mode)
+         assert_equal(0, File.umask & File.stat(@dir2).mode)
       end
    end
 
@@ -35,8 +40,8 @@ class TC_Dir_Mkdir_Class < Test::Unit::TestCase
    end
 
    def teardown
-      remove_dir(@dir1)
-      remove_dir(@dir2)
+      remove_dir(@dir1) if File.exists?(@dir1)
+      remove_dir(@dir2) if File.exists?(@dir2)
 
       @dir1 = nil
       @dir2 = nil
